feat: getting song recommendations for a playlist, and user can add them
