+++ src/components/RecommendationSongs/RecommendationSongs.css
@@ -1,0 +1,4 @@
+.RecommendationSongs {
+  width: 100%;
+  height: 100%;
+}
