+++ src/components/RecommendationSongs/RecommendationSongs.tsx
@@ -9,21 +9,21 @@ import { SpotifyTrackType } from '../../types/SpotifyTrackType';
 
 // Interface
 interface Props {
-  songs: Array<{
-    track: SpotifyTrackType
-  }>
+  songs: Array<SpotifyTrackType>
+  roomCode: string
 }
 
-const PlaylistSongs: React.FC<Props> = (props) => {
+const RecommendationSongs: React.FC<Props> = (props) => {
   return (
-    <div className='PlaylistSongs'>
-      <h2>Playlist Songs:</h2>
+    <div className='RecommendationSongs'>
+      <h2>Recommendations:</h2>
       {props.songs && props.songs.map((item, index) => {
         return (
           <Song
             key={index}
-            song={item.track}
-            addSong={false}
+            song={item}
+            room={props.roomCode}
+            addSong={true}
           />
         )
       })}
@@ -31,4 +31,4 @@ const PlaylistSongs: React.FC<Props> = (props) => {
   )
 }
 
-export default PlaylistSongs
+export default RecommendationSongs
